+++ client/web/src/routes.tsx
@@ -194,7 +194,8 @@ export const routes: readonly LayoutRouteProps<any>[] = [
         render: lazyComponent(() => import('./insights/InsightsPage'), 'InsightsPage'),
         condition: props =>
             !isErrorLike(props.settingsCascade.final) &&
-            !!props.settingsCascade.final?.experimentalFeatures?.codeInsights,
+            !!props.settingsCascade.final?.experimentalFeatures?.codeInsights &&
+            props.settingsCascade.final['insights.displayLocation.insightsPage'] !== false,
     },
     {
         path: '/views',